Make folders and separate carrot shooting logic
