+++ Assets/Scenes/BossStage1Scene.unity
@@ -11242,6 +11242,7 @@ GameObject:
   - component: {fileID: 1474742876}
   - component: {fileID: 1474742877}
   - component: {fileID: 1474742878}
+  - component: {fileID: 1474742880}
   - component: {fileID: 1474742879}
   m_Layer: 0
   m_Name: BossStageManager
@@ -11345,6 +11346,22 @@ MonoBehaviour:
   areaMax: {x: 10, y: 1.5, z: 10}
   warningDuration: 1
   dropHeight: 15
+--- !u!114 &1474742880
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1474742874}
+  m_Enabled: 0
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0943a64c1ae0946e98a4bfadba8af765, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  animator: {fileID: 2039046444}
+  carrotSpawnOffset: {fileID: 2039046445}
+  carrotPf: {fileID: 5649900439809378948, guid: 1eecc96bfe961436396ca41d4ed49458, type: 3}
+  carrotDelayTime: 1.6
 --- !u!1001 &1495427503
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -13724,6 +13741,16 @@ PrefabInstance:
       objectReference: {fileID: 2534964839176971238, guid: 169224d39859543e19c5277c072237d7, type: 3}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 2aa73c6bbedca41769b8a7d9e6630c26, type: 3}
+--- !u!95 &2039046444 stripped
+Animator:
+  m_CorrespondingSourceObject: {fileID: 367596013406301719, guid: 2aa73c6bbedca41769b8a7d9e6630c26, type: 3}
+  m_PrefabInstance: {fileID: 2039046443}
+  m_PrefabAsset: {fileID: 0}
+--- !u!4 &2039046445 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 367596013244964437, guid: 2aa73c6bbedca41769b8a7d9e6630c26, type: 3}
+  m_PrefabInstance: {fileID: 2039046443}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1001 &2045720548
 PrefabInstance:
   m_ObjectHideFlags: 0
