+++ Assets/Scripts/Boss/CarrotAttack.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5d89403f851a14964957c92f3c061b98
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
