+++ Assets/Scripts/Boss/CarrotAttack/CarrotManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0943a64c1ae0946e98a4bfadba8af765
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
