+++ Assets/Scripts/Boss/WeakSpot.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3e3247d9ae0ab41b7beb7416d54afc64
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
